# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "description": "The one and only error response shape in this API.",
        "examples": [
          {
            "error": {
              "code": "forbidden",
              "message": "You do not have access to this resource."
            }
          },
          {
            "error": {
              "code": "validation_error",
              "details": {
                "fields": [
                  {
                    "field": "body.password",
                    "reason": "String should have at least 12 characters",
                    "type": "string_too_short"
                  }
                ]
              },
              "message": "The request failed validation."
            }
          }
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ProvisionAdminRequest": {
        "additionalProperties": false,
        "description": "Create an admin account. Admin-only; there is no self-service path.",
        "examples": [
          {
            "email": "ops@saci.example",
            "password": "correct-horse-battery-staple"
          }
        ],
        "properties": {
          "email": {
            "examples": [
              "founder@example.com"
            ],
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Initial password, at least 12 characters. The new admin must verify their email and enrol MFA before any admin capability opens to them.",
            "maxLength": 256,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "ProvisionAdminRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "description": "Access token issued by `POST /v1/auth/login`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for the FundReady platform. All resources live under `/v1`. Timestamps are ISO 8601 UTC, currencies are ISO 4217 codes, and money is expressed in integer minor units.\n\nEvery error response uses the envelope `{\"error\": {\"code\", \"message\", \"details\"}}`. Branch on `code`, which is stable; never parse `message`.\n\nEvery response carries an `X-Request-ID` header. Quote it in a support report and we can find the exact log line.",
    "summary": "AI business audit and investor-matching platform for SACI Holdings.",
    "title": "FundReady API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/users": {
      "post": {
        "description": "Creates an admin. This is the only way one comes into existence -- `/v1/auth/register` refuses the role.\n\nThe new admin lands `pending_verification` and without MFA, so they must verify their address and enrol a second factor before any admin capability opens to them. A verification email is sent.\n\nUnlike self-registration, a duplicate address returns `409` rather than a uniform success: the caller is already trusted, so there is no enumeration concern.\n\nRequires an admin account **with MFA enrolled**. An admin who has not enrolled receives `403` here until they do.",
        "operationId": "provision_admin_v1_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionAdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Provision an admin account",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/admin/users/{user_id}/reactivate": {
      "post": {
        "description": "Returns the account to `active`, or to `pending_verification` if the address was never verified -- reactivating does not confer verification.\n\n`409` if the account is not suspended.\n\nRequires an admin account **with MFA enrolled**. An admin who has not enrolled receives `403` here until they do.",
        "operationId": "reactivate_user_v1_admin_users__user_id__reactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Lift a suspension",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/admin/users/{user_id}/role": {
      "patch": {
        "description": "Applies the new role and **forces the user to log in again**: `session_valid_after` moves, so their existing access and refresh tokens stop working and the next login is authorised against the new role.\n\n`403` if you target your own account. `409` if the user already has that role, or if demoting them would leave no active admin.\n\nRequires an admin account **with MFA enrolled**. An admin who has not enrolled receives `403` here until they do.",
        "operationId": "change_user_role_v1_admin_users__user_id__role_patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Change a user's role",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/admin/users/{user_id}/suspend": {
      "post": {
        "description": "Blocks the account and **ends its sessions immediately** -- refresh tokens revoked, access tokens already issued invalidated.\n\n`403` if you target your own account. `409` if the target is the only active admin, since the only way back from zero admins is database credentials.\n\nRequires an admin account **with MFA enrolled**. An admin who has not enrolled receives `403` here until they do.",
        "operationId": "suspend_user_v1_admin_users__user_id__suspend_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Suspend an account",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "description": "Exchanges email and password for tokens.\n\n**Branch on `status`.** `authenticated` returns `tokens`. `mfa_required` means the password was correct but a second factor is outstanding: send the returned `mfa_token` and the code to `POST /v1/auth/mfa/verify`. The `mfa_token` grants nothing by itself and expires in 5 minutes.\n\nEvery failure returns the same `401` with the same message -- whether the account does not exist, the password is wrong, or the account is temporarily locked after repeated failures. Do not branch on it.\n\nA `403` means the account is suspended. Logging in while `pending_verification` succeeds, but protected endpoints return `403` until the email is verified.",
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "summary": "Log in",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "description": "Revokes the presented refresh token and every token from the same login. The access token expires on its own shortly after.\n\nAlways returns `204`, even for a token we do not recognise.",
        "operationId": "logout_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "summary": "Log out",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}